backend/requirements.txt
django-ninja

flower
gunicorn
docker
celery[pytest,redis]
cryptography
pycryptodome
redis
Jinja2